requirements.txt
Fabric==1.6.1
distribute==0.6.31
ipython==0.13.2
paramiko==1.10.1
pycrypto==2.6
python-vagrant==0.3.1
wsgiref==0.1.2
